backend/tour_api/requirements.txt
flask
flask-cors
gunicorn
mysql-connector-python
python-dotenv
requests

